Update Nomad Client/Server RPC codecs to use custom msgpackHandle
